feat: créditos fracionados na API (float) e formatação pt-BR no painel

+++ apps/api/app/schemas/admin_dashboard.py
@@ -16,14 +16,14 @@ class NewTenantsPerDay(BaseModel):
 
 
 class CreditsSummary(BaseModel):
-    sold: int
-    consumed: int
+    sold: float
+    consumed: float
 
 
 class LowBalanceTenant(BaseModel):
     id: uuid.UUID
     name: str
-    credit_balance: int
+    credit_balance: float
 
 
 class WhatsappConnectedSummary(BaseModel):
